feat: Update MQTT configuration and add settings screen
This commit updates the MQTT configuration in the `config.dart` file, changing the topic to 'notificationTopic'. It also adds a new `SettingsScreen` class that allows the user to configure the MQTT IP address. The IP address is saved using the `shared_preferences` package and can be accessed later for MQTT connection.