feat: Add Flutter Local Notifications package and implement test notification
This commit adds the `flutter_local_notifications` package to the project and implements a test notification feature. The `showTestNotification` method is added to the `_DashboardScreenState` class in the `dashboard_screen.dart` file. This method displays a test notification with a title and body when the user taps on the notification add button. The notification is shown using the `flutterLocalNotificationsPlugin` instance and the `AndroidNotificationDetails` and `NotificationDetails` classes. This feature enhances the user experience by allowing them to test the notification functionality within the app.