feat: Update notification title and body in dashboard_screen.dart
This commit updates the notification title and body in the `_DashboardScreenState` class of the `dashboard_screen.dart` file. The title is changed to "Aviso sobre seus locais" and the body is changed to "Houve um problema no seu local. Abra o app para maior informação". This improves the clarity and relevance of the notification message for the user.